% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracting_digits.R
\name{first_digit}
\alias{first_digit}
\title{Extract the first digit from a numeric vector}
\usage{
first_digit(numbers)
}
\arguments{
\item{numbers}{numeric}
}
\value{
numeric
}
\description{
Extract the first digit from a numeric vector
}
\examples{
first_digit(c(123, 0.09, -678.9))
}
