% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracting_digits.R
\name{first_digit_}
\alias{first_digit_}
\title{Function to extract first two digits}
\usage{
first_digit_(number)
}
\arguments{
\item{number}{numeric}
}
\value{
character
}
\description{
Function to extract first two digits
}
