% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracting_digits.R
\name{first_two_digits}
\alias{first_two_digits}
\title{Extract the first two digits from a numeric vector}
\usage{
first_two_digits(numbers)
}
\arguments{
\item{numbers}{Numeric vector}
}
\value{
Numeric vector
}
\description{
Extract the first two digits from a numeric vector
}
\examples{
first_two_digits(c(123, 0.09, -678.9))
}
