% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracting_digits.R
\name{last_digit}
\alias{last_digit}
\title{Extract the last digit from a numeric vector}
\usage{
last_digit(numbers)
}
\arguments{
\item{numbers}{Numeric Vector}
}
\value{
numeric vector
}
\description{
Extract the last digit from a numeric vector
}
\examples{
last_digit(c(12, 3.141, 2.781, 0.998))
}
