% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracting_digits.R
\name{last_two_digits}
\alias{last_two_digits}
\title{Extract the last two digits from a numeric vector}
\usage{
last_two_digits(numbers)
}
\arguments{
\item{numbers}{Numeric Vector}
}
\value{
numeric vector
}
\description{
Extract the last two digits from a numeric vector
}
