% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digit_probability.R
\name{prob_first_digit}
\alias{prob_first_digit}
\title{Calculate the probability of a first digit according to Benford's Law.}
\usage{
prob_first_digit(digit)
}
\arguments{
\item{digit}{An integer ranging from 1-9.}
}
\value{

}
\description{
Calculate the probability of a first digit according to Benford's Law.
}
\examples{
# to get Benford's law expected digit distribution:
prob_first_digit(1:9)
}
