% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digit_probability.R
\name{prob_first_two_digits}
\alias{prob_first_two_digits}
\title{Calculate the probability of the first two digits according to Benford's Law.}
\usage{
prob_first_two_digits(digits)
}
\arguments{
\item{digits}{First two digits of the number.}
}
\value{

}
\description{
Calculate the probability of the first two digits according to Benford's Law.
}
\examples{
# What is the probability of two ones
prob_first_two_digits(11)
}
