% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digit_probability.R
\name{prob_second_digit}
\alias{prob_second_digit}
\title{Calculate the probabilty of a second digit.}
\usage{
prob_second_digit(d2)
}
\arguments{
\item{d2}{Integer or vector of integers that presents the second digit.}
}
\value{
Double or Numeric vector
}
\description{
Can be used to detect bias in number reporting.
For example, when rounding up is occuring, numbers with the
second digit as zero will be overrepresented, while
higher numbers like (7,8,9) will be underrepresented
in the sample of numbers.
}
\examples{
# calculate probabilty of second digits
prob_second_digit(0:9)

# cumulative probabilty must be equal to one
sum(prob_second_digit(0:9))

}
