% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second_order_test.R
\name{second_order_transf}
\alias{second_order_transf}
\title{Second-Order Transformation to create Benford Set}
\usage{
second_order_transf(numbers)
}
\arguments{
\item{numbers}{Numeric Vector}
}
\value{
Numeric vector
}
\description{
From some data distributions it is expected that they don't follow
Benford's Law. In such cases it is annying that you can't use Benford's
Law to evaluate the data. However, after ordering and taking the differences
between these numbers, the transformed data should conform to Benford's Law.
}
\details{
This transformation can be applied to Normal, Uniform, Gamma and Triangular Distributions.
}
\examples{
# generate some numbers
numbers <- rnorm(1000)

# assure data complies to a Benford Set
tranformed_numbers <- second_order_transf(numbers)

# Now the data must conform to Benford's Law
table(first_digit(tranformed_numbers))
}
