% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracting_digits.R
\name{to_scientific}
\alias{to_scientific}
\title{Convert numeric to scientific notation}
\usage{
to_scientific(numbers)
}
\arguments{
\item{numbers}{numeric.}
}
\value{
character
}
\description{
Convert numeric to scientific notation
}
\examples{
to_scientific(c(12.112,  435, -0.09))
}
