% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracting_digits.R
\name{extract_leading_digits}
\alias{extract_leading_digits}
\title{Extract and return leading digits as characters}
\usage{
extract_leading_digits(numbers)
}
\arguments{
\item{numbers}{numeric}
}
\value{
character
}
\description{
Extract and return leading digits as characters
}
